% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Downloads detailed COVID-19 case data from default BCCDC webpage URL
(passed in as string) as a .csv file to default relative path directory
(also passed in as string).}
\usage{
get_data()
}
\arguments{
\item{url}{string A webpage from which .csv file is downloaded to relative
data folder directory, provided wit ha default value.}

\item{out_folder}{string A relative path directory comprising two folder
separated by a forward slash ("/"), provided with a default value.}
}
\value{
A data.frame containing downloaded detailed case data.
}
\description{
Downloads detailed COVID-19 case data from default BCCDC webpage URL
(passed in as string) as a .csv file to default relative path directory
(also passed in as string).
}
\examples{
get_data()
get_data(url = "http://www.bccdc.ca/Health-Info-Site/Documents/
  BCCDC_COVID19_Dashboard_Case_Details.csv", out = "data/raw")
}
