% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_by_cond.R
\name{plot_hist_by_cond}
\alias{plot_hist_by_cond}
\title{Plot histogram of Covid19 cases}
\usage{
plot_hist_by_cond(startDate, endDate, condition)
}
\arguments{
\item{startDate}{string the start date of the period (no earlier than '2020-01-29')}

\item{endDate}{string the end date of the period (no later than today)}

\item{condition}{string available condition - Age, Region}
}
\value{
ggplot object
}
\description{
Plots the number of Covid19 cases with respect to the condition(per Age Group
or Region) using histogram for the period specified by startDate and endDate
}
\examples{
plot_hist_by_cond("2021-01-01", "2021-12-31", "Age")
plot_hist_by_cond("2021-01-01", "2021-12-31", "Region")
}
