% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_by_date.R
\name{plot_line_by_date}
\alias{plot_line_by_date}
\title{Line plot of Covid19 cases}
\usage{
plot_line_by_date(startDate, endDate, region = "all")
}
\arguments{
\item{startDate}{string the start date of the period (no earlier than '2020-01-29')}

\item{endDate}{string the end date of the period (no later than today)}

\item{region}{vector or string = 'all' Default value is string 'all' - displaying all regions.
Other available values: combination of list of strings from available regions -
Fraser, Vancouver Coastal, Vancouver Island, Interior, Northern, Out of Canada}
}
\value{
ggplot object
}
\description{
Plots the line chart of regional Covid19 cases over the period specified by
startDate and endDate (format: YYYY-MM-DD). The default argument value
for region is 'all', showing the total number of Covid19 cases in BC.
}
\examples{
plot_line_by_date("2021-01-01", "2021-12-31")
plot_line_by_date("2021-01-01", "2021-12-31", region = c('Fraser'))
}
