% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_summary_stat.r
\name{show_summary_stat}
\alias{show_summary_stat}
\title{Show summary statistics of BC covid data}
\usage{
show_summary_stat(startDate, endDate)
}
\arguments{
\item{startDate}{string  The start date of the period (no earlier than '2020-01-29').}

\item{endDate}{string The end date of the period (no later than today).}
}
\value{
A tibble containing summary statistics with the following columns:
total_cases_count, latest_date, latest_daily_cases_count,
max_date, max_daily_cases_count, min_date, min_daily_cases_count,
max_age_group, max_age_group_count, min_age_group, min_age_group_count,
max_region, max_region_count, min_region, min_region_count
}
\description{
Show summary statistics for the Covid19 cases in BC
for the period specified by startDate and endDate (format: YYYY-MM-DD).
}
\examples{
show_summary_stat("2022-01-01", "2022-01-13")

}
